Solidification of targeting mode handling
Targeting modes now inherit TargetingMode class and are stored in target_selection folder.

Database scans that folder and stores the scripts in a globally accesible array

Level ui now has "right panel" which will be expanded in the future, for now it gets all targeting modes and allows them to be selected.

-renamed TargetSelection to TargetingMode
-base_turret perform_targeting() now uses TargetingMode.select()